# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Type of input provided
 */
export type GenerateInputInputType = typeof GenerateInputInputType[keyof typeof GenerateInputInputType];


export const GenerateInputInputType = {
  url: 'url',
  text: 'text',
} as const;

export interface GenerateInput {
  /** YouTube URL or raw text idea */
  input: string;
  /** Type of input provided */
  inputType?: GenerateInputInputType;
}

export interface Script {
  /** Viral opening hook */
  hook: string;
  /** Full script body with visual cues */
  body: string;
  /** Recommended hashtags */
  hashtags: string[];
  /**
     * Visual direction notes
     * @nullable
     */
  visualCues?: string | null;
}

export interface GenerateResult {
  /** Unique ID for this generation */
  id: string;
  scripts: Script[];
  /**
     * Summary of the input provided
     * @nullable
     */
  inputSummary?: string | null;
}

export interface TranscriptInput {
  /** YouTube video URL */
  url: string;
}

export interface TranscriptResult {
  /** Extracted transcript text */
  transcript: string;
  /**
     * Video title if available
     * @nullable
     */
  title?: string | null;
}

export interface ScriptRecord {
  id: string;
  createdAt: string;
  inputSummary: string;
  scripts: Script[];
}

export interface AuthInput {
  email: string;
  password: string;
}

export interface AuthResult {
  token: string;
  userId: string;
}

export interface ErrorResponse {
  error: string;
}