# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthInput {
  email: string;
  password: string;
}