# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateInputInputType } from './generateInputInputType';

export interface GenerateInput {
  /** YouTube URL or raw text idea */
  input: string;
  /** Type of input provided */
  inputType?: GenerateInputInputType;
}