# lib/api-zod/src/generated/types/generateResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Script } from './script';

export interface GenerateResult {
  /** Unique ID for this generation */
  id: string;
  scripts: Script[];
  /**
     * Summary of the input provided
     * @nullable
     */
  inputSummary?: string | null;
}