# lib/api-zod/src/generated/types/script.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Script {
  /** Viral opening hook */
  hook: string;
  /** Full script body with visual cues */
  body: string;
  /** Recommended hashtags */
  hashtags: string[];
  /**
     * Visual direction notes
     * @nullable
     */
  visualCues?: string | null;
}