# lib/api-zod/src/generated/types/scriptRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Script } from './script';

export interface ScriptRecord {
  id: string;
  createdAt: Date;
  inputSummary: string;
  scripts: Script[];
}