# lib/api-zod/src/generated/types/transcriptResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Retno API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptResult {
  /** Extracted transcript text */
  transcript: string;
  /**
     * Video title if available
     * @nullable
     */
  title?: string | null;
}